1) Bugfix: bowtie2 doesn't handle spaces in the index path very well, have to add a backslash along with quotation marks. 2) Optimization: cutadapt now throws out reads of length 0 and 1, as those are ignored by bowtie2 and emit a warning. 3) Added an option to remove universal Illumina adapters. 4) Added better control over which files are removed and which are kept.

+++ MapFastq.py
@@ -1,14 +1,10 @@
-import glob, os
-import zlib
-import gzip
-from Bio.Seq import Seq
-from Bio.Alphabet import generic_dna
-import subprocess
+import os
 from subprocess import PIPE, Popen
 CutAdaptCMD = '/usr/local/bin/cutadapt'
 TnPrimerAndTail = 'GTATTTTACCGACCGTTACCGACCGTTTTCATCCCTA'
 BowtiePath = '/usr/local/bin/'
-CInd = 'dependencies/albicans/reference genome/C_albicans_SC5314_version_A22-s07-m01-r08_chromosomes_HapA'
+# NB: bowtie2 requires spaces to be escapes with a backslash for the -x parameter.
+CInd = 'dependencies/albicans/reference\ genome/C_albicans_SC5314_version_A22-s07-m01-r08_chromosomes_HapA'
 CORES = 4 # How many cores on the machine = how many threads should the external tools utilize
 
 def GetReads(Val,Text):
@@ -35,7 +31,7 @@ def cmdline(command):
 def RemoveTn(InputFileName, TnSeq, overlap, OutputFName):
     #-g is from the beggining and -a is from the end or vice versa (if we use -a we get only a string length 6, which is probably the 6 Ns between the primer and Tn)
     #-o out put, afterwards the input. --overlap determines the minimal number of bases found from the Tn
-    cmdl = CutAdaptCMD + ' -g ' + TnSeq + ' -o "'+ OutputFName + '" "' + InputFileName + '" --discard-untrimmed --overlap ' + str(overlap)
+    cmdl = CutAdaptCMD + ' -m 2 -g ' + TnSeq + ' -o "'+ OutputFName + '" "' + InputFileName + '" --discard-untrimmed --overlap ' + str(overlap)
     Output,err = cmdline(cmdl)
     return ProcessOutput(Output)
     
@@ -44,6 +40,9 @@ def RemoveTn(InputFileName, TnSeq, overlap, OutputFName):
 usage = """USAGE: MapFastq.py  
    -o  --OutDir          [str]   Output fir, if not specified creating all file in curr dir.
    -i  --InputFileName   [str]   Input fastq file name 
+   -a  --CleanAdapters           Clean Illumina universl adapters.
+   -d  --DeleteOriginals         Delete input FASTQ files.
+   -k  --KeepCleanFqs            Keep the cleaned FASTQ files. 
    -h, --help                    print this help 
 """
 
@@ -51,14 +50,20 @@ if __name__ == '__main__':
     import sys
     import getopt
     
-    OutputDir = ""; FastqFName = ""; 
+    OutputDir = ""
+    FastqFName = ""
+    clean_adapters = False
+    delete_originals = False
+    keep_clean_fqs = False
     
     try:                                
-        opts, args = getopt.getopt(sys.argv[1:], "o:i:h", ["OutDir=","InputFileName=","help"])
+        opts, args = getopt.getopt(sys.argv[1:], "o:i:adkh",
+                                   ["OutDir=","InputFileName=","CleanAdapters",
+                                    "DeleteOriginals","KeepCleanFqs","help"])
     except getopt.GetoptError:
         print usage
-        sys.exit(2)                     
-    for opt, arg in opts:                
+        sys.exit(2)
+    for opt, arg in opts:
         if opt in ("-h", "--help"):      
             print usage
             sys.exit()                  
@@ -66,6 +71,12 @@ if __name__ == '__main__':
             OutputDir = arg
         elif opt in ("-i", "--InputFileName"): 
             FastqFName = arg
+        elif opt in ("-a", "--CleanAdapters"):
+            clean_adapters = True
+        elif opt in ("-d", "--DeleteOriginals"):
+            delete_originals = True
+        elif opt in ("-k", "--KeepCleanFqs"):
+            keep_clean_fqs = True
 
     if  len(FastqFName)==0 or len(CInd) ==0:
         print "Input file or index file not specified, existing."
@@ -74,11 +85,17 @@ if __name__ == '__main__':
 
     CleanfName = os.path.join(OutputDir, os.path.basename(FastqFName) + '.clean.fq')
     CurrRes = RemoveTn(FastqFName,TnPrimerAndTail,37, CleanfName)
+    if clean_adapters:
+        temp_fq = CleanfName + ".no_adapters.fq"
+        cmdline(r'''%s -m 2 -a AGATCGGAAGAGCACACGTCTGAACTCCAGTCAC -o "%s" "%s"''' %
+                (CutAdaptCMD, temp_fq, CleanfName))
+        os.unlink(CleanfName)
+        os.rename(temp_fq, CleanfName)
     #aligning fastq file to the reference genome
     #-x reference_genome; -U fq file of unpaired reads; -S output SAM alignment file; X is the max fragment size to consider (relevant only in paired end)
     SamfName = os.path.join(OutputDir, os.path.splitext(os.path.basename(FastqFName))[0] + '.sam')
-    BowtieCmd = BowtiePath + 'bowtie2 -p ' + str(CORES) + ' -x '+ CInd + \
-                ' -U "' + CleanfName + '" -S "' + SamfName + '"'
+    BowtieCmd = BowtiePath + 'bowtie2 -p ' + str(CORES) + ' -x "'+ CInd + \
+                '" -U "' + CleanfName + '" -S "' + SamfName + '"'
     Output,err = cmdline(BowtieCmd)
     # converting to bam file 
     cmdline('samtools view -@ ' + str(CORES) + ' -bS "'+ SamfName + '" > "' +
@@ -90,8 +107,10 @@ if __name__ == '__main__':
     cmdline('samtools index "' + os.path.splitext(SamfName)[0] + '.sorted.bam"')
 
     # Remove intermediats:
-    # os.remove(FastqFName) # By default, don't remove the original FASTQ.
-    os.remove(CleanfName)
+    if delete_originals:
+        os.remove(FastqFName)
+    if not keep_clean_fqs:
+        os.remove(CleanfName)
     os.remove(SamfName)
     os.remove(os.path.splitext(SamfName)[0] + '.bam')
